refactor: Move transparency helper structs out of `api/` directory
Keep API about API logic. Export helper structs through API exporter. Allow
isolated unit tests for helper structs.
