+++ neptune-core/src/models/blockchain/transaction/mod.rs
@@ -15,6 +15,8 @@ pub mod lock_script;
 pub mod primitive_witness;
 pub mod transaction_kernel;
 pub mod transaction_proof;
+pub mod transparent_input;
+pub mod transparent_transaction_details;
 pub mod utxo;
 pub(crate) mod utxo_triple;
 pub mod validity;
