+++ neptune-core/src/models/blockchain/transaction/transparent_input.rs
